made minor based on errors in compiling, code now compiles
