+++ Exchange.py
@@ -1,13 +1,10 @@
 # Neo operations
 from boa.builtins import concat
-from boa.interop.Neo.Runtime import GetTime
 
 # Contract constants
-from Constants import bucket_duration, max_fee
+from Constants import max_fee
 # Storage Manager
-from MCTManager import put, get, deserialize, serialize_array
-# Offer actions
-from Offer import set_volume
+from MCTManager import put, get
 
 
 # Owner functions
@@ -66,10 +63,6 @@ def initialize():
     put('state', 'Active')
 
 
-def get_state():
-    return get('state')
-
-
 def get_maker_fee(asset_id):
     key = concat("makerFee", asset_id)
     fee = get(key)
@@ -94,44 +87,3 @@ def get_balance(originator, asset_id):
     key = concat(originator, asset_id)
     return get(key)
 
-
-def get_exchange_rate(asset_id):
-    time = GetTime()
-
-    bucket_number = time / bucket_duration
-
-    return get_volume(bucket_number, asset_id)
-
-
-def get_volume(bucket_number, asset_id):
-    volume_key = concat("tradeVolume", bucket_number)
-    volume_key = concat(volume_key, asset_id)
-    volume_data = get(volume_key)
-
-    if len(volume_data) == 0:
-        return set_volume()
-    else:
-        return deserialize(volume_data)
-
-
-def add_volume(asset_id, native_amount, foreign_amount):
-    time = GetTime()
-
-    bucket_number = time / bucket_duration
-
-    volume_key = concat("tradeVolume", bucket_number)
-    volume_key = concat(volume_key, asset_id)
-
-    volume_data = get(volume_key)
-
-    if len(volume_data) == 0:
-        volume = set_volume()
-
-        volume["Native"] = native_amount
-        volume["Foreign"] = foreign_amount
-    else:
-        volume = deserialize(volume_data)
-        volume["Native"] = volume["Native"] + native_amount
-        volume["Foreign"] = volume["Foreign"] + foreign_amount
-
-    put(volume_key, serialize_array(volume))
